+++ Fun.Result/Option.fs
@@ -23,7 +23,16 @@ module OptionComputationExpression =
     type OptionBuilder() =
         member inline _.Return(x) = Some x
         member inline _.ReturnFrom(x) = x
+
         member inline _.Bind(x, [<InlineIfLambda>] f) = Option.bind f x
+
+        member inline _.Bind(x: Result<_, _>, [<InlineIfLambda>] f) =
+            Option.bind
+                f
+                (match x with
+                 | Ok x -> Some x
+                 | _ -> None)
+
         member inline _.Delay([<InlineIfLambda>] f) = f ()
         member inline _.Zero() = Some()