+++ Fun.Result/TaskResult.fs
@@ -10,21 +10,21 @@ type TaskResult<'Success, 'Failure> = Task<Result<'Success, 'Failure>>
 
 [<RequireQualifiedAccess>]
 module TaskResult =
-    let map f (x: TaskResult<_, _>) : TaskResult<_, _> = Task.map (Result.map f) x
+    let inline map f (x: TaskResult<_, _>) : TaskResult<_, _> = Task.map (Result.map f) x
 
-    let mapError f (x: TaskResult<_, _>) : TaskResult<_, _> = Task.map (Result.mapError f) x
+    let inline mapError f (x: TaskResult<_, _>) : TaskResult<_, _> = Task.map (Result.mapError f) x
 
-    let retn x : TaskResult<_, _> = Ok x |> Task.FromResult
+    let inline retn x : TaskResult<_, _> = Ok x |> Task.FromResult
 
 
-    let bind (f: _ -> TaskResult<_, _>) (x: TaskResult<_, _>) : TaskResult<_, _> = task {
+    let inline bind (f: _ -> TaskResult<_, _>) (x: TaskResult<_, _>) : TaskResult<_, _> = task {
         let! result = x
         match result with
         | Ok x -> return! f x
         | Error x -> return (Error x)
     }
 
-    let bindError (f: _ -> Task<_>) (x: TaskResult<_, _>) : TaskResult<_, _> = task {
+    let inline bindError (f: _ -> Task<_>) (x: TaskResult<_, _>) : TaskResult<_, _> = task {
         let! result = x
         match result with
         | Ok x -> return Ok x
@@ -33,37 +33,40 @@ module TaskResult =
             return (Error newErr)
     }
 
-    let ofSuccess x : TaskResult<_, _> = Ok x |> Task.FromResult
+    let inline ofSuccess x : TaskResult<_, _> = Ok x |> Task.FromResult
 
-    let ofError x : TaskResult<_, _> = Error x |> Task.FromResult
+    let inline ofError x : TaskResult<_, _> = Error x |> Task.FromResult
 
-    let ofAsync (x: Async<_>) : TaskResult<_, _> = task {
+    let inline ofAsync (x: Async<_>) : TaskResult<_, _> = task {
         let! result = x
         return Ok result
     }
 
-    let ofTask (x: Task<_>) : TaskResult<_, _> = x |> Task.map Ok
+    let inline ofTask (x: Task<_>) : TaskResult<_, _> = x |> Task.map Ok
 
-    let ofEmptyTask (x: Task) : TaskResult<_, _> = task {
+    let inline ofEmptyTask (x: Task) : TaskResult<_, _> = task {
         do! x
         return Ok()
     }
 
-    let bindTask f x = x |> ofTask |> bind f
+    let inline bindTask f x = x |> ofTask |> bind f
 
-    let mapTask f x = x |> ofTask |> map f
+    let inline mapTask f x = x |> ofTask |> map f
 
 
 [<AutoOpen>]
 module TaskResultComputationExpression =
     type TaskResultBuilder() =
-        member __.Return(x) = TaskResult.retn x
-        member __.Bind(x: TaskResult<_, _>, f) = TaskResult.bind f x
-        member __.ReturnFrom(x) = x
-        member __.Delay(f) = f
-        member __.Run(f) = f ()
+        member inline __.Return(x) = TaskResult.retn x
 
-        member this.Zero() = this.Return()
+        member inline __.Bind(x: TaskResult<_, _>, [<InlineIfLambda>] f) = TaskResult.bind f x
+        member inline __.Bind(x: Result<_, _>, [<InlineIfLambda>] f) = TaskResult.bind f (Task.retn x)
+
+        member inline __.ReturnFrom(x) = x
+        member inline __.Delay(f) = f
+        member inline __.Run([<InlineIfLambda>] f) = f ()
+
+        member inline this.Zero() = this.Return()
 
         member this.While(guard, body: unit -> TaskResult<_, _>) =
             if not (guard ()) then
@@ -71,21 +74,21 @@ module TaskResultComputationExpression =
             else
                 this.Bind(body (), (fun () -> this.While(guard, body)))
 
-        member this.TryWith(body, handler) = task {
+        member inline this.TryWith(body, handler) = task {
             try
                 return! this.ReturnFrom(body ())
             with e ->
                 return! handler e
         }
 
-        member this.TryFinally(body, compensation) = task {
+        member inline this.TryFinally(body, compensation) = task {
             try
                 return! this.ReturnFrom(body ())
             finally
                 compensation ()
         }
 
-        member this.Using(disposable: #System.IDisposable, body) =
+        member inline this.Using(disposable: #System.IDisposable, body) =
             let body' = fun () -> body disposable
             this.TryFinally(
                 body',
@@ -95,9 +98,9 @@ module TaskResultComputationExpression =
                     | disp -> disp.Dispose()
             )
 
-        member this.For(sequence: seq<_>, body) =
+        member inline this.For(sequence: seq<_>, body) =
             this.Using(sequence.GetEnumerator(), (fun enum -> this.While(enum.MoveNext, this.Delay(fun () -> body enum.Current))))
-        member this.Combine(a: TaskResult<_, _>, b) = this.Bind(a, (fun () -> b ()))
+        member inline this.Combine(a: TaskResult<_, _>, b) = this.Bind(a, (fun () -> b ()))
 
     let taskResult = TaskResultBuilder()