+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [2.1.0] - 2024-06-05
+
+- Add more inline to improve performance
+- Support let! Result type in TaskResult
+
 ## [2.0.9] - 2023-03-17
 
 - Fix try with/finally for asyncOption and asyncResultOption